Fix add products x market mock.

+++ Backend/src/main/java/com/mobile/ComparaCompra/domain/ProductXMarket.java
@@ -14,7 +14,7 @@ public class ProductXMarket
 {
     Product product;
     long idMarket;
-    long price;
+    double price;
     String imageUrl;
     double distance;
 }
